Move GFA-related functions to separate source file
